requirements.txt
openai
rq
nltk
transformers
flask
gunicorn
python-dotenv
pinecone-client
psycopg2-binary
